Expression is not an integer constant expression
Fix "Expression is not an integer constant expression"-error on iOS 8 Beta 3

+++ Masonry/MASUtilities.h
@@ -15,11 +15,11 @@
     #define MASEdgeInsets UIEdgeInsets
 
     enum {
-        MASLayoutPriorityRequired = UILayoutPriorityRequired,
-        MASLayoutPriorityDefaultHigh = UILayoutPriorityDefaultHigh,
+        MASLayoutPriorityRequired = (int)UILayoutPriorityRequired,
+        MASLayoutPriorityDefaultHigh = (int)UILayoutPriorityDefaultHigh,
         MASLayoutPriorityDefaultMedium = 500,
-        MASLayoutPriorityDefaultLow = UILayoutPriorityDefaultLow,
-        MASLayoutPriorityFittingSizeLevel = UILayoutPriorityFittingSizeLevel,
+        MASLayoutPriorityDefaultLow = (int)UILayoutPriorityDefaultLow,
+        MASLayoutPriorityFittingSizeLevel = (int)UILayoutPriorityFittingSizeLevel,
     };
     typedef float MASLayoutPriority;
